% File library/man/dh.Rd
\name{dh}
\alias{dh}
\title{dh}
\description{
The function changes the last generation in the population to double haploid
}
\usage{
dh(pop)
}
\arguments{
\item{pop}{Required: an input population}}
\value{
A haplotypes' object
}
\examples{

pop2 = dh(pop1)}


