% File library/man/draw_population.Rd
\name{draw_population}
\alias{draw_population}
\title{draw_population}
\description{
The function plots haplotypes
}
\usage{
draw_population(pop,parental_genotypes,file,im_type)
}
\arguments{
\item{pop}{Required: a haplotype object input.}
\item{parental_genotypes}{Required: the parental genotyping object}
\item{file}{Required: the output file}
\item{im_type}{Optional: the type of images, default is svg}}
\examples{
draw_population(pop,parental_genotypes,file)}
