% File library/man/haplotypes_in.Rd
\name{haplotypes_in}
\alias{haplotypes_in}
\title{haplotypes_in}
\description{
The function restores a saved genotypes
}
\usage{
haplotypes_in(input)
}
\arguments{
\item{input}{The prefix of the input file.}}
\value{
A haplotypes' object from a file
}
\examples{

haplotypes = haplotypes_in('file_haplo')}