% File library/man/pick_individual.Rd
\name{pick_individual}
\alias{pick_individual}
\title{pick_individual}
\description{
The function picks an individual from a population
}
\usage{
pick_individual(pop,cross,family,individual)
}
\arguments{
\item{pop}{Required: an input population}
\item{cross}{Optional: the cross level to be select from}
\item{family}{Optional: the family level to be select from}
\item{individual}{Required: an individual to be selected}}
\examples{

pick_individual('input_script')}
