% File library/man/run_pipeline.Rd
\name{run_pipeline}
\alias{run_pipeline}
\title{run_pipeline}
\description{
The function runs a full pipeline in crossword syntax supplied in an input file.}
\usage{
run_pipeline(script_file)
}
\arguments{
\item{script_file}{The prefix of the input file.}}
\examples{
script_file = "crossword_script_input_file.script"
run_pipeline('input_script')}
