% File library/man/add_header_button.Rd
\name{add_header_button}
\alias{add_header_button}
\title{add_header_button}
\description{
The function create a dialog box of headers
}
\usage{
add_header_button(pop,output,parental_genotypes)
}
\arguments{
\item{e}{Required: gedit object to receive the output}}
\examples{
add_header_button(gedit())}
